move test that was misplaced in #5990
